chore: restore swift examples
